# cmd/rogerai-broker/openapi.yaml
openapi: 3.1.0
info:
  title: RogerAI Broker API
  version: 0.1.0
  description: |
    The RogerAI broker - the only public component of the RogerAI marketplace.
    It is an OpenAI-compatible relay plus a registry, wallet, and metering layer
    over a fleet of crowd-sourced, self-hosted LLM nodes.

    Roles:
      * Consumers call `POST /v1/chat/completions` (OpenAI-compatible) and browse
        `GET /discover`. Spending and wallet reads are SIGNED with a per-user
        Ed25519 key (headers `X-Roger-Pubkey` / `X-Roger-TS` / `X-Roger-Sig`); the
        broker derives a stable wallet id from the verified pubkey. The legacy
        `X-Roger-User` header is accepted only as an UNAUTHENTICATED hint and can
        never spend a signed user's wallet. (See the `signedUser` security scheme.)
      * Providers ("nodes") dial OUT: `POST /nodes/register`, then long-poll
        `GET /agent/poll` and `POST /agent/result`. No inbound ports.
      * Owners who MONETIZE log in with GitHub (`POST /auth/github` from the CLI
        device flow, or the `/auth/github/login` web flow) to bind their account.

    Every served request yields a hash-chained UsageReceipt signed by the node and
    counter-signed by the broker (the "lineage" guarantee).
  license: { name: Proprietary }
servers:
  - url: https://broker.rogerai.fyi
  - url: http://127.0.0.1:7070

tags:
  - { name: meta, description: Service metadata }
  - { name: consumer, description: Discover + use models }
  - { name: billing, description: Prepaid credit wallet (Stripe) }
  - { name: auth, description: GitHub login (owners who monetize) }
  - { name: provider, description: Node registration + the outbound job tunnel }

paths:
  /health:
    get:
      tags: [meta]
      summary: Liveness probe
      responses: { "200": { description: ok, content: { text/plain: { schema: { type: string, example: ok } } } } }

  /:
    get:
      tags: [meta]
      summary: Service descriptor
      responses:
        "200":
          description: Service info
          content: { application/json: { schema: { $ref: '#/components/schemas/ServiceInfo' } } }

  /openapi.yaml:
    get:
      tags: [meta]
      summary: This OpenAPI document
      responses: { "200": { description: spec, content: { application/yaml: {} } } }

  /discover:
    get:
      tags: [consumer]
      summary: List model offers (cheapest active price first)
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                type: object
                properties:
                  offers: { type: array, items: { $ref: '#/components/schemas/Offer' } }

  /balance:
    get:
      tags: [consumer]
      summary: Wallet balance for the caller
      parameters:
        - $ref: '#/components/parameters/Pubkey'
        - $ref: '#/components/parameters/SigTS'
        - $ref: '#/components/parameters/Sig'
        - $ref: '#/components/parameters/User'
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                type: object
                properties: { user: { type: string }, balance: { type: number } }

  /me:
    get:
      tags: [consumer]
      summary: Consumer dashboard - balance, lifetime spend, recent requests
      description: |
        The caller's spend dashboard: current wallet balance, lifetime total spend
        (sum of settled costs), and the most-recent settled requests (newest first).
      parameters:
        - $ref: '#/components/parameters/Pubkey'
        - $ref: '#/components/parameters/SigTS'
        - $ref: '#/components/parameters/Sig'
        - $ref: '#/components/parameters/User'
        - $ref: '#/components/parameters/Limit'
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { type: string }
                  balance: { type: number, description: current wallet credits }
                  spend: { type: number, description: lifetime credits spent }
                  recent: { type: array, items: { $ref: '#/components/schemas/Entry' } }

  /earnings:
    get:
      tags: [provider]
      summary: Owner dashboard - accrued earnings + recent requests for a node
      description: |
        A node owner's earnings dashboard: accrued (unpaid) owner credits for the
        node and its most-recent settled requests (newest first). The node id is the
        source of truth - a serving node knows its own id.
      parameters:
        - { in: query, name: node, required: true, schema: { type: string }, description: the node id }
        - $ref: '#/components/parameters/Limit'
      responses:
        "200":
          description: Earnings
          content:
            application/json:
              schema:
                type: object
                properties:
                  node: { type: string }
                  online: { type: boolean }
                  earnings: { type: number, description: accrued (unpaid) owner credits }
                  recent: { type: array, items: { $ref: '#/components/schemas/Entry' } }
        "400": { description: node query param required }

  /market:
    get:
      tags: [consumer]
      summary: Per-model market metrics (supply, load, speed, quality, signal)
      description: |
        A live, per-model marketplace view aggregated from node state: how many
        providers are online, current in-flight load, the cheapest active input
        price, the best measured throughput, the mean success rate, and a 0..100
        "signal" combining supply, quality, and reliability (lightly discounted by
        congestion). Sorted by signal, highest first. The signal is a glanceable
        health bar, NOT a price.
      responses:
        "200":
          description: Market
          content:
            application/json:
              schema:
                type: object
                properties:
                  market: { type: array, items: { $ref: '#/components/schemas/MarketModel' } }

  /v1/chat/completions:
    post:
      tags: [consumer]
      summary: OpenAI-compatible chat completion, relayed to the best node
      description: |
        Drop-in OpenAI endpoint. The broker matches an online node serving `model`
        (cheapest active price; optionally filtered by the constraint headers),
        relays the request, meters it, and settles the wallet. The price the user
        is first quoted for a (node, model) is honored for 24h (published
        time-of-use / free windows are charged as-is). Set `stream:true` for SSE.
      security: [ { signedUser: [] } ]
      parameters:
        - $ref: '#/components/parameters/Pubkey'
        - $ref: '#/components/parameters/SigTS'
        - $ref: '#/components/parameters/Sig'
        - $ref: '#/components/parameters/User'
        - { in: header, name: X-Roger-Confidential, schema: { type: string }, description: "If set, route only to TEE-attested confidential nodes." }
        - { in: header, name: X-Roger-Min-TPS, schema: { type: number }, description: "Require measured throughput >= this (tokens/sec)." }
        - { in: header, name: X-Roger-Max-Price, schema: { type: number }, description: "Skip nodes whose active input price exceeds this (credits/1M)." }
        - { in: header, name: X-Roger-Node, schema: { type: string }, description: "Pin to a specific node (used by client-side failover to target a chosen provider)." }
        - { in: header, name: X-Roger-Exclude-Nodes, schema: { type: string }, description: "Comma-separated node ids to skip (the providers a client just saw fail) - lets the connector route AROUND a dropped provider." }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChatRequest' } } }
      responses:
        "200":
          description: |
            Completion (verbatim upstream body). Response headers carry the meter:
            `X-RogerAI-Provider`, `X-RogerAI-Cost`, `X-RogerAI-Balance`,
            `X-RogerAI-Price`, `X-RogerAI-TPS`, `X-RogerAI-Receipt` (the signed receipt).
          content: { application/json: { schema: { type: object } } }
        "401": { description: invalid request signature, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "402": { description: insufficient credits, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "503": { description: no node offers the model (or matches constraints), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "504": { description: node timed out }

  /billing/checkout:
    post:
      tags: [billing]
      summary: Create a Stripe Checkout session to buy credits
      parameters: [ { $ref: '#/components/parameters/User' } ]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { usd: { type: number, default: 10 } } } } }
      responses:
        "200":
          description: Checkout link
          content:
            application/json:
              schema: { type: object, properties: { url: { type: string }, usd: { type: number }, credits: { type: number } } }
        "503": { description: billing not configured }

  /billing/webhook:
    post:
      tags: [billing]
      summary: Stripe webhook (checkout.session.completed → credit wallet)
      description: HMAC-verified via `Stripe-Signature`; idempotent (credits each session once).
      parameters: [ { in: header, name: Stripe-Signature, required: true, schema: { type: string } } ]
      responses: { "200": { description: received }, "400": { description: bad signature } }

  /auth/github:
    post:
      tags: [auth]
      summary: Bind a GitHub owner to the caller's signing pubkey (login to monetize)
      description: |
        Login-to-monetize. After a GitHub OAuth Device Flow (`rogerai login`), the
        CLI posts its GitHub access token here over a SIGNED request. The broker
        verifies the token against `GET https://api.github.com/user` server-side,
        then binds `github_id` <-> `login` <-> the signing pubkey as an owner.
        Earning operations (priced node registration, future withdraws) require this
        binding; the consume/free paths never do. The access token is verified and
        discarded - never stored or logged.
      security: [ { signedUser: [] } ]
      parameters:
        - $ref: '#/components/parameters/Pubkey'
        - $ref: '#/components/parameters/SigTS'
        - $ref: '#/components/parameters/Sig'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [access_token], properties: { access_token: { type: string, description: GitHub OAuth access token from the device flow } } }
      responses:
        "200":
          description: Bound
          content:
            application/json:
              schema: { type: object, properties: { ok: { type: boolean }, github_login: { type: string }, github_id: { type: integer } } }
        "400": { description: access_token required }
        "401": { description: unsigned request or GitHub token rejected, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  /nodes/register:
    post:
      tags: [provider]
      summary: Register/refresh a provider node and its offers
      description: |
        Proof-of-possession: signed with the node key (anti-replay TS). A node
        advertising a NONZERO price is an EARNING node and the request must ALSO be
        signed with the owner's user key (the `signedUser` scheme) bound to a GitHub
        owner via POST /auth/github - otherwise 401 (unsigned) / 403 (not linked).
        Free (all-zero-price) nodes register with no login.
      parameters:
        - $ref: '#/components/parameters/Pubkey'
        - $ref: '#/components/parameters/SigTS'
        - $ref: '#/components/parameters/Sig'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/NodeRegistration' } } }
      responses:
        "200": { description: ok }
        "401": { description: invalid signature, or priced offer without a signed owner request }
        "403": { description: priced offer but the signing key is not a GitHub-linked owner }

  /nodes/heartbeat:
    post:
      tags: [provider]
      summary: Keep a node marked online
      requestBody:
        content: { application/json: { schema: { type: object, properties: { node_id: { type: string } } } } }
      responses: { "200": { description: ok } }

  /agent/poll:
    get:
      tags: [provider]
      summary: Long-poll for a relayed job (held up to 25s)
      security: [ { nodeToken: [] } ]
      parameters: [ { in: query, name: node, required: true, schema: { type: string } } ]
      responses:
        "200": { description: a job, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
        "204": { description: no job within the window - re-poll }
        "401": { description: bad node token }

  /agent/result:
    post:
      tags: [provider]
      summary: Return a served job's result + signed receipt
      security: [ { nodeToken: [] } ]
      parameters: [ { in: query, name: node, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/JobResult' } } }
      responses: { "200": { description: ok }, "401": { description: bad node token } }

  /agent/stream:
    post:
      tags: [provider]
      summary: Pipe a streaming job's SSE chunks back to the client
      description: For `stream:true` jobs, the node POSTs the upstream SSE body here and the broker forwards it, flushing, to the waiting consumer; the node then posts the receipt to /agent/result.
      security: [ { nodeToken: [] } ]
      parameters:
        - { in: query, name: node, required: true, schema: { type: string } }
        - { in: query, name: job, required: true, schema: { type: string } }
      responses: { "200": { description: piped }, "401": { description: bad node token }, "404": { description: no active stream } }

components:
  securitySchemes:
    nodeToken:
      type: http
      scheme: bearer
      description: The per-node BridgeToken set at registration.
    signedUser:
      type: apiKey
      in: header
      name: X-Roger-Sig
      description: |
        Consumer request signature. The local proxy signs every broker request with
        the user's Ed25519 key and sends three headers: `X-Roger-Pubkey` (hex public
        key), `X-Roger-TS` (unix seconds, must be within 5 minutes), and
        `X-Roger-Sig` (hex ed25519 signature over the canonical string
        `method + "\n" + path + "\n" + ts + "\n" + hex(sha256(body))`). The broker
        derives the wallet id as `u_` + first 16 hex of sha256(pubkey). A present but
        invalid signature is rejected with 401.
  parameters:
    User:
      in: header
      name: X-Roger-User
      schema: { type: string }
      description: |
        Legacy, UNAUTHENTICATED caller hint. Honored only when no signature is
        present; cannot spend or read a signed user's wallet. Prefer the signed
        headers (see the `signedUser` security scheme).
    Pubkey:
      in: header
      name: X-Roger-Pubkey
      schema: { type: string }
      description: Hex ed25519 public key of the signing user (part of the request signature).
    SigTS:
      in: header
      name: X-Roger-TS
      schema: { type: integer }
      description: Unix-seconds timestamp the signature covers (must be within 5 minutes).
    Sig:
      in: header
      name: X-Roger-Sig
      schema: { type: string }
      description: Hex ed25519 signature over the canonical request string.
    Limit:
      in: query
      name: limit
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      description: Max history rows to return (default 20, clamped to 100).
  schemas:
    ServiceInfo:
      type: object
      properties:
        service: { type: string, example: rogerai-broker }
        version: { type: string }
        spec: { type: string, example: /openapi.yaml }
    Error:
      type: object
      properties: { error: { type: object, properties: { message: { type: string } } } }
    ChatRequest:
      type: object
      required: [model, messages]
      properties:
        model: { type: string }
        messages: { type: array, items: { type: object } }
        stream: { type: boolean }
        max_tokens: { type: integer }
    PriceWindow:
      type: object
      description: Time-of-use rule (HH:MM UTC; may wrap midnight; empty days = daily).
      properties:
        days: { type: array, items: { type: integer } }
        start: { type: string, example: "18:00" }
        end: { type: string, example: "22:00" }
        price_in: { type: number }
        price_out: { type: number }
        free: { type: boolean }
    ModelOffer:
      type: object
      properties:
        model: { type: string }
        price_in: { type: number, description: credits per 1M input tokens (base/fallback) }
        price_out: { type: number, description: credits per 1M output tokens (base/fallback) }
        ctx: { type: integer }
        schedule: { type: array, items: { $ref: '#/components/schemas/PriceWindow' } }
    Offer:
      allOf:
        - $ref: '#/components/schemas/ModelOffer'
        - type: object
          properties:
            node_id: { type: string }
            region: { type: string }
            hw: { type: string }
            online: { type: boolean }
            confidential: { type: boolean }
            free_now: { type: boolean }
            scheduled: { type: boolean }
            tps: { type: number, description: measured output tokens/sec (0 = unmeasured) }
    NodeRegistration:
      type: object
      required: [node_id, pub_key, offers]
      properties:
        node_id: { type: string }
        pub_key: { type: string, description: hex ed25519 public key }
        bridge_token: { type: string, description: shared secret authenticating this node's poll/result }
        region: { type: string }
        hw: { type: string }
        offers: { type: array, items: { $ref: '#/components/schemas/ModelOffer' } }
        confidential: { type: boolean }
        attestation: { type: string, description: TEE attestation blob (verified before the confidential badge) }
    UsageReceipt:
      type: object
      description: Hash-chained, node-signed + broker-counter-signed per-request record.
      properties:
        request_id: { type: string }
        node_id: { type: string }
        user: { type: string, description: pseudonymized per (user, node) }
        model: { type: string }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        price_in: { type: number }
        price_out: { type: number }
        ts: { type: integer }
        prev_hash: { type: string }
        lineage_method: { type: string }
        node_sig: { type: string }
        broker_sig: { type: string }
    MarketModel:
      type: object
      description: Aggregated market metrics for one model (see GET /market).
      properties:
        model: { type: string }
        providers: { type: integer, description: online nodes offering this model }
        in_flight: { type: integer, description: active requests across those nodes }
        min_price: { type: number, description: cheapest active input price (credits/1M) }
        best_tps: { type: number, description: fastest measured output tok/s }
        success_rate: { type: number, description: mean EWMA success across providers (0..1) }
        signal: { type: integer, description: 0..100 demand/quality/reliability signal }
    Entry:
      type: object
      description: One settled request, as surfaced to the /me and /earnings dashboards.
      properties:
        request_id: { type: string }
        user: { type: string, description: real wallet owner (dashboards are owner-private) }
        node: { type: string }
        model: { type: string }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        cost: { type: number, description: credits the consumer paid }
        owner_share: { type: number, description: credits credited to the node owner }
        ts: { type: integer }
    Job:
      type: object
      properties:
        id: { type: string }
        user: { type: string }
        body: { type: object, description: the raw OpenAI request }
    JobResult:
      type: object
      properties:
        id: { type: string }
        status: { type: integer }
        body: { type: object }
        receipt: { $ref: '#/components/schemas/UsageReceipt' }